Backend/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.23.1
binaryornot==0.4.4
black==24.2.0
bleach==6.1.0
blinker==1.7.0
boto3==1.34.36
botocore==1.34.36
build==1.1.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cookiecutter==2.6.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.3
docopt==0.6.2
entrypoints==0.4
executing==2.0.0
faiss-cpu==1.8.0
fastapi==0.96.1
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.16.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.4.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
keras==2.15.0
kiwisolver==1.4.5
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.26
libclang==16.0.6
Mako==1.3.2
mangum==0.17.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.1
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pickleshare==0.7.5
pillow==10.2.0
pipdeptree==2.16.1
pipreqs==0.4.13
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-json-logger==2.0.7
python-multipart==0.0.5
python-slugify==8.0.4
python-socketio==5.11.1
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
redis==4.6.0
referencing==0.30.2
reflex==0.4.4
reflex-hosting-cli==0.1.9
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.12.0
rsa==4.9
ruff==0.2.1
s3transfer==0.10.0
scikit-learn==1.2.2
scipy==1.12.0
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlmodel==0.0.14
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
starlette-admin==0.9.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tinycss2==1.2.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.12.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.20.0
watchdog==2.3.1
watchfiles==0.19.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
